Fix all current issue + C# 9 + remove usseless delay (#110)
* Remove useless delay

- This was causing issue about player not receiving it's item when escaping

* Fix Output position

* 914 Fix

- Fix Ammo not being clear
- optimise it to keep item in inventory

* C#9

* Fix #108

* Fix Bad Yamato moment

* Remove C# 9 change

* Plugin should be load in priority

- Make than it's don't overwrite value of other plugins

* Fix i don't want this file

* Fix Error

* ItemChance::Chance change to float instead of int

* More configurable Random

* FixWarhead

* Oups

* API update

+++ Common Utilities/ConfigObjects/ItemUpgradeChance.cs
@@ -4,9 +4,9 @@ namespace Common_Utilities.ConfigObjects
     {
         public ItemType Original { get; set; }
         public ItemType New { get; set; }
-        public int Chance { get; set; }
+        public double Chance { get; set; }
 
-        public void Deconstruct(out ItemType itemType, out ItemType itemType1, out int i)
+        public void Deconstruct(out ItemType itemType, out ItemType itemType1, out double i)
         {
             itemType = Original;
             itemType1 = New;
