+++ Common Utilities/ConfigObjects/Scp914EffectChance.cs
@@ -5,10 +5,10 @@ namespace Common_Utilities.ConfigObjects
     public class Scp914EffectChance
     {
         public EffectType Effect { get; set; }
-        public int Chance { get; set; }
+        public double Chance { get; set; }
         public float Duration { get; set; }
 
-        public void Deconstruct(out EffectType effect, out int chance, out float duration)
+        public void Deconstruct(out EffectType effect, out double chance, out float duration)
         {
             effect = Effect;
             chance = Chance;
