+++ Common Utilities/ConfigObjects/Scp914TeleportChance.cs
@@ -8,10 +8,10 @@ namespace Common_Utilities.ConfigObjects
         public ZoneType Zone { get; set; } = ZoneType.Unspecified;
         public RoomType Room { get; set; }
         public Vector3 Offset { get; set; } = Vector3.zero;
-        public int Chance { get; set; }
+        public double Chance { get; set; }
         public float Damage { get; set; } = 0f;
 
-        public void Deconstruct(out RoomType room, out Vector3 offset, out int chance, out float damage, out ZoneType zone)
+        public void Deconstruct(out RoomType room, out Vector3 offset, out double chance, out float damage, out ZoneType zone)
         {
             room = Room;
             offset = Offset;
